+++ ballcat-notify/ballcat-notify-biz/src/main/java/com/hccake/ballcat/notify/recipient/SpecifyOrganizationRecipientFilter.java
@@ -36,7 +36,7 @@ public class SpecifyOrganizationRecipientFilter implements RecipientFilter {
 	 */
 	@Override
 	public List<SysUser> filter(List<Object> filterCondition) {
-		List<Integer> organizationIds = filterCondition.stream().map(Integer.class::cast).collect(Collectors.toList());
+		List<Long> organizationIds = filterCondition.stream().map(Long.class::cast).collect(Collectors.toList());
 		return sysUserService.listByOrganizationIds(organizationIds);
 	}
