+++ ballcat-notify/ballcat-notify-controller/src/main/java/com/hccake/ballcat/notify/controller/UserAnnouncementController.java
@@ -49,7 +49,7 @@ public class UserAnnouncementController {
 	@PreAuthorize("@per.hasPermission('notify:userannouncement:read')")
 	@Operation(summary = "用户公告已读上报", description = "用户公告已读上报")
 	public R<Void> readAnnouncement(@PathVariable("announcementId") Long announcementId) {
-		Integer userId = SecurityUtils.getUser().getUserId();
+		Long userId = SecurityUtils.getUser().getUserId();
 		userAnnouncementService.readAnnouncement(userId, announcementId);
 		return R.ok();
 	}
